Move awesome_overrides files to concerns
