+++ lib/decidim/decidim_awesome/engine.rb
@@ -36,7 +36,12 @@ module Decidim
         end
 
         Decidim::Proposals::ApplicationHelper.include(Decidim::DecidimAwesome::Proposals::ApplicationHelperOverride)
+        Decidim::Proposals::ProposalWizardCreateStepForm.include(Decidim::DecidimAwesome::Proposals::ProposalWizardCreateStepFormOverride)
+
         Decidim::AmendmentsHelper.include(Decidim::DecidimAwesome::AmendmentsHelperOverride)
+
+        Decidim::MenuPresenter.include(Decidim::DecidimAwesome::MenuPresenterOverride)
+        Decidim::MenuItemPresenter.include(Decidim::DecidimAwesome::MenuItemPresenterOverride)
       end
 
       initializer "decidim_awesome.view_helpers" do
